+++ src/Plainion.GraphViz.Infrastructure/Properties/AssemblyInfo.cs
@@ -6,15 +6,15 @@ using System.Reflection;
 [assembly: AssemblyProduct("Plainion.GraphViz")]
 [assembly: AssemblyDescription("Plainion.GraphViz")]
 [assembly: AssemblyCopyright("Copyright @ 2022")]
-[assembly: AssemblyVersion("4.2")]
-[assembly: AssemblyFileVersion("4.2")]
+[assembly: AssemblyVersion("4.3")]
+[assembly: AssemblyFileVersion("4.3")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyTitle = "Plainion.GraphViz.Infrastructure";
         internal const System.String AssemblyProduct = "Plainion.GraphViz";
         internal const System.String AssemblyDescription = "Plainion.GraphViz";
         internal const System.String AssemblyCopyright = "Copyright @ 2022";
-        internal const System.String AssemblyVersion = "4.2";
-        internal const System.String AssemblyFileVersion = "4.2";
+        internal const System.String AssemblyVersion = "4.3";
+        internal const System.String AssemblyFileVersion = "4.3";
     }
 }